+++ src/components/maple/result/characterInfo/CharacterInfo.tsx
